Major changes to the virtual page handling. Virtual has been renamed to Pattern. Pattern pages have been created for all the core custom pages Archive Taxonomy Contact Search Single Update
for both Jobs and Experts

Pages are initially created with a default set of button and a shortcode for the page in question, Example:

<p style="text-align: center;">[jbp-expert-post-btn][jbp-job-post-btn][jbp-expert-browse-btn][jbp-job-browse-btn][jbp-expert-profile-btn][jbp-job-list-btn]</p>
[jbp-job-update-page]

Users can edit these pages adding text or other shortcodes around the page shortcodes. Title and slugs can be edited, though that is only really useful for the update pages because archive titles are created by wordpress and single pages are whatever theuser ienters on an actual listing.



git-svn-id: https://aux1.edublogs.org/repos/incsub/trunk/plugins/jobs-plus@14830 f53fe46e-8c02-459b-9fdd-93e75771b278