Makes some changes to allow Open-ended Jobs.
It makes the Completion Date field not "required" and adds an N/A notice on Due date if nothing is inputted.

It filters the existing array and adds an open-ended option to avoid hiccups when updating the plugin.

It states the job as Open-ended instead of "Expired" if openended is selected.

Please see: https://app.asana.com/0/11289012348292/1109524727319922

It includes screenshots as well and proper scope for testing.